Now the user should be able to choose which one to use with os.environ["NENGIN_BACKEND"] = "opengl"
